refactor(docker-package): move shell interface in its file
